+++ docs/source/conf.py
@@ -106,6 +106,7 @@ extensions = [
 
 # Set that source code from plotting is always included
 plot_include_source = True
+plot_html_show_source_link = True
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
@@ -301,7 +302,6 @@ def package_list_from_file(file):
 # define mapping from PyPI names to python imports
 PACKAGE_MAPPING = {
     "PyYAML": "yaml",
-    "pytorch-lightning": "pytorch_lightning",
 }
 MOCK_PACKAGES = []
 if SPHINX_MOCK_REQUIREMENTS: