+++ src/torchmetrics/functional/image/ssim.py
@@ -72,7 +72,7 @@ def _ssim_update(
             The luminance term can be obtained with luminance=ssim/contrast
             Mutually exclusive with ``return_full_image``
     """
-    is_3d = len(preds.shape) == 5
+    is_3d = preds.ndim == 5
 
     if not isinstance(kernel_size, Sequence):
         kernel_size = 3 * [kernel_size] if is_3d else 2 * [kernel_size]
@@ -367,7 +367,7 @@ def _multiscale_ssim_update(
     """
     mcs_list: List[Tensor] = []
 
-    is_3d = len(preds.shape) == 5
+    is_3d = preds.ndim == 5
 
     if not isinstance(kernel_size, Sequence):
         kernel_size = 3 * [kernel_size] if is_3d else 2 * [kernel_size]