+++ tests/integrations/test_lightning.py
@@ -14,13 +14,17 @@
 from unittest import mock
 
 import torch
-from pytorch_lightning import LightningModule, Trainer
+from lightning_utilities import module_available
 from torch import tensor
 from torch.nn import Linear
-from torch.utils.data import DataLoader
+
+if module_available("lightning"):
+    from lightning import LightningModule, Trainer
+else:
+    from pytorch_lightning import LightningModule, Trainer
 
 from integrations.helpers import no_warning_call
-from integrations.lightning.boring_model import BoringModel, RandomDataset
+from integrations.lightning.boring_model import BoringModel
 from torchmetrics import MetricCollection, SumMetric
 from torchmetrics.classification import BinaryAccuracy, BinaryAveragePrecision
 
@@ -40,7 +44,7 @@ def test_metric_lightning(tmpdir):
         def __init__(self) -> None:
             super().__init__()
             self.metric = SumMetric()
-            self.sum = 0.0
+            self.register_buffer("sum", torch.tensor(0.0))
 
         def training_step(self, batch, batch_idx):
             x = batch
@@ -176,11 +180,11 @@ def test_metric_lightning_log(tmpdir):
             super().__init__()
             self.metric_step = SumMetric()
             self.metric_epoch = SumMetric()
-            self.sum = torch.tensor(0.0)
+            self.register_buffer("sum", torch.tensor(0.0))
             self.outs = []
 
         def on_train_epoch_start(self):
-            self.sum = torch.tensor(0.0)
+            self.sum = torch.tensor(0.0, device=self.sum.device)
 
         def training_step(self, batch, batch_idx):
             x = batch
@@ -221,8 +225,8 @@ def test_metric_collection_lightning_log(tmpdir):
         def __init__(self) -> None:
             super().__init__()
             self.metric = MetricCollection([SumMetric(), DiffMetric()])
-            self.sum = torch.tensor(0.0)
-            self.diff = torch.tensor(0.0)
+            self.register_buffer("sum", torch.tensor(0.0))
+            self.register_buffer("diff", torch.tensor(0.0))
 
         def training_step(self, batch, batch_idx):
             x = batch
@@ -265,7 +269,7 @@ def test_scriptable(tmpdir):
             # the metric is not used in the module's `forward`
             # so the module should be exportable to TorchScript
             self.metric = SumMetric()
-            self.sum = torch.tensor(0.0)
+            self.register_buffer("sum", torch.tensor(0.0))
 
         def training_step(self, batch, batch_idx):
             x = batch