+++ tests/unittests/deprecations/root_class_imports.py
@@ -4,11 +4,45 @@ from functools import partial
 import pytest
 
 from torchmetrics import (
+    BLEUScore,
+    CharErrorRate,
+    CHRFScore,
+    ErrorRelativeGlobalDimensionlessSynthesis,
+    ExtendedEditDistance,
+    MatchErrorRate,
+    ModifiedPanopticQuality,
+    MultiScaleStructuralSimilarityIndexMeasure,
+    PanopticQuality,
+    PeakSignalNoiseRatio,
     PermutationInvariantTraining,
+    Perplexity,
+    RelativeAverageSpectralError,
+    RetrievalFallOut,
+    RetrievalHitRate,
+    RetrievalMAP,
+    RetrievalMRR,
+    RetrievalNormalizedDCG,
+    RetrievalPrecision,
+    RetrievalPrecisionRecallCurve,
+    RetrievalRecall,
+    RetrievalRecallAtFixedPrecision,
+    RetrievalRPrecision,
+    RootMeanSquaredErrorUsingSlidingWindow,
+    SacreBLEUScore,
     ScaleInvariantSignalDistortionRatio,
     ScaleInvariantSignalNoiseRatio,
     SignalDistortionRatio,
     SignalNoiseRatio,
+    SpectralAngleMapper,
+    SpectralDistortionIndex,
+    SQuAD,
+    StructuralSimilarityIndexMeasure,
+    TotalVariation,
+    TranslationEditRate,
+    UniversalImageQualityIndex,
+    WordErrorRate,
+    WordInfoLost,
+    WordInfoPreserved,
 )
 from torchmetrics.functional.audio import scale_invariant_signal_noise_ratio
 
@@ -16,6 +50,7 @@ from torchmetrics.functional.audio import scale_invariant_signal_noise_ratio
 @pytest.mark.parametrize(
     "metric_cls",
     [
+        # Audio
         pytest.param(
             partial(PermutationInvariantTraining, scale_invariant_signal_noise_ratio), id="PermutationInvariantTraining"
         ),
@@ -23,6 +58,44 @@ from torchmetrics.functional.audio import scale_invariant_signal_noise_ratio
         ScaleInvariantSignalNoiseRatio,
         SignalDistortionRatio,
         SignalNoiseRatio,
+        # Detection
+        ModifiedPanopticQuality,
+        PanopticQuality,
+        # Image
+        ErrorRelativeGlobalDimensionlessSynthesis,
+        MultiScaleStructuralSimilarityIndexMeasure,
+        PeakSignalNoiseRatio,
+        RelativeAverageSpectralError,
+        RootMeanSquaredErrorUsingSlidingWindow,
+        SpectralAngleMapper,
+        SpectralDistortionIndex,
+        StructuralSimilarityIndexMeasure,
+        TotalVariation,
+        UniversalImageQualityIndex,
+        # Info Retrieval
+        RetrievalFallOut,
+        RetrievalHitRate,
+        RetrievalMAP,
+        RetrievalMRR,
+        RetrievalNormalizedDCG,
+        RetrievalPrecision,
+        RetrievalPrecisionRecallCurve,
+        RetrievalRecall,
+        RetrievalRecallAtFixedPrecision,
+        RetrievalRPrecision,
+        # Text
+        BLEUScore,
+        CharErrorRate,
+        CHRFScore,
+        ExtendedEditDistance,
+        MatchErrorRate,
+        Perplexity,
+        SacreBLEUScore,
+        SQuAD,
+        TranslationEditRate,
+        WordErrorRate,
+        WordInfoLost,
+        WordInfoPreserved,
     ],
 )
 def test_import_from_root_package(metric_cls):