+++ tests/unittests/image/test_ms_ssim.py
@@ -100,9 +100,9 @@ class TestMultiScaleStructuralSimilarityIndexMeasure(MetricTester):
 
 def test_ms_ssim_contrast_sensitivity():
     """Test that the contrast sensitivity is correctly computed with 3d input."""
-    preds = torch.rand(1, 1, 182, 182, 182)
-    target = torch.rand(1, 1, 182, 182, 182)
+    preds = torch.rand(1, 1, 50, 50, 50)
+    target = torch.rand(1, 1, 50, 50, 50)
     out = multiscale_structural_similarity_index_measure(
-        preds, target, data_range=1.0, kernel_size=11, betas=(1.0, 0.5, 0.25)
+        preds, target, data_range=1.0, kernel_size=3, betas=(1.0, 0.5, 0.25)
     )
     assert isinstance(out, torch.Tensor)